Fix receipts failing to open in Firefox
Firefox blocks extensions from opening data URLs which we were using.
A workaround has been added that opens a ZRA page and injects a form into that instead of loading it using a data URL.

To make sure this fix is only applied in firefox, a new environment variable, BROWSER, has been added.