Fix tabLoaded not resolving when tabs are closed quickly
The fix for fast loading receipt tabs added in 6c28792eee52a3f17b1671093b18a52fe72cb4a4 meant that the listener for tab close events was added too late. Thus, if a tab was closed shortly after it was opened, there was a chance that event would be missed.