Add support for single HTML file downloading
This is an alternative to MHTML files, which can only be downloaded using the Chrome extensions API. Thus, actions which download pages are now supported in Firefox as well as Chrome.

The single HTML files are essentially just HTML files with all resource URLs replaced with data URLs. This excludes scripts which are stripped from the file.

Since downloading these single HTML files is faster, HTML is now the default download format but MHTML files can still be downloaded using a setting.

Despite being faster, downloading these single HTML files causes some performance issues. Many HTTP requests are made and despite the results being cached, the extension still slows down when many pages are being downloaded in parallel. In the future, a better caching system should be implemented that deals with the case where multiple requests are made to the same resource at the same time and thus can't use the cache as none of the requests have completed. Until then, to help mitigate this and other issues, new settings were added to limit the number of parallel requests, parallel downloads and to configure the delay between downloads. All this is done using the new PromiseQueue class which handles the concurrency limiting as well as delaying.