Fix login feature closing the logged in tab if any errors occur
Login was originally designed to close the tab if errors were thrown when getting the tab ID. This was because nothing else would have access to the tab ID and would thus be unable to close the tab.

Since the login only feature doesn't actually do anything with the opened tab or its ID, this autoclosing behaviour can be disabled.

Fixes #128