Fix system error detection and output
To detect common system errors, we must check the original return that new payments this week belong to. Problem was that those returns were not linked to the payments since they were from the previous week. This fixes that links returns to payments twice. Once to determine unbalanced records and again for use when checking for system errors.

Additionally, system errors are now attached to the correct records when generating reason strings. E.g. if an 'unallocated advance payment' error is detected, the error will be attached to the advance payment that caused it, and not the return that was used to detect it.

Other changes:
- Allow records to have more than one system error
- Add basic system for tracking non-critical errors during processing that should be shown to the user
- Update more JSDoc of functions that support any record that extends the base record
- Allow 'unallocated advance payment' error to occur even if there was also a 'return rounded up' error
- Handle closing balance not existing