Standardize properties and methods related to retrying in action runners
- Add failures prop
- Add helper methods
- Remove runAllTaxTypes from return history runner since it's basically just runInternal now
- Remove setRetryReason

The new methods are mostly borrowed from what was already added to the return history client action such as, `getRetryReasons` and `getRetryInput`. They help to make runners more flexible by making it easier to override retry reasons and inputs in sub-classes.

Of note is a change to the behaviour of retrying; the retry reason and input of runners are now only set when `checkIfShouldRetry` is called.