Update ZRA website domain
ZRA recently changed their website URL from https://www.zra.org.zm to https://eservices.zra.org.zm:8000.

Note that the port in the new URL is not included in the manifest.json because Firefox does not accept it. When it is included, Firefox starts throwing network errors even if though the same manifest works fine in Chrome.