Add functions to allow getting the full version of any page of a report
The existing functions only got the internal table of the reports.

To get a full report page, the first page is first retrieved in full and then if the desired page is not the first page, the minimal inner table version of the page is retrieved and then merged with the first full page.